requirements.txt
ultralytics>=8.3.0
opencv-python>=4.5.0
flask>=2.0.0
pillow>=8.3.0
numpy>=1.19.0
pyyaml>=5.4.0
pycocotools>=2.0.0
